Update the AWS provider to 3.75+
The 3.75.0 release of the Terraform AWS provider backports changes made
to the configuration of S3 buckets in the 4.0 release of the provider.
This will allow us to make the necessary changes to prepare for
upgrading to v4 of the AWS provider. As a result of this update, the
configuration for encrypting the buckets is moved out into their own
resources.